Moved Cloudflare method unit tests.
