+++ tests/Unit/Addons/Cloudflare/Cloudflare/setCacheLevel.php
@@ -5,7 +5,12 @@ use WP_Rocket\Tests\Unit\TestCase;
 use WP_Rocket\Addons\Cloudflare\Cloudflare;
 use Brain\Monkey\Functions;
 
-class TestHasPageRule extends TestCase {
+/**
+ * @covers WP_Rocket\Addons\Cloudflare\Cloudflare::set_cache_level
+ *
+ * @group Cloudflare
+ */
+class Test_SetCacheLevel extends TestCase {
 
 	protected function setUp() {
 		parent::setUp();
@@ -22,9 +27,9 @@ class TestHasPageRule extends TestCase {
 	}
 
 	/**
-	 * Test Cloudflare has page rules with cached invalid transient.
+	 * Test set cache level with cached invalid transient.
 	 */
-	public function testHasRuleWithInvalidCredentials() {
+	public function testSetCacheLevelWithInvalidCredentials() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
@@ -38,22 +43,20 @@ class TestHasPageRule extends TestCase {
 		$cloudflare_facade_mock->shouldNotReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
 
 		$this->assertEquals(
 			$wp_error,
-			$has_page_rule
+			$cloudflare->set_cache_level( 'aggressive' )
 		);
 	}
 
 	/**
-	 * Test Cloudflare has page rules with exception.
+	 * Test set cache level with exception.
 	 */
-	public function testHasRuleWithException() {
+	public function testSetCacheLevelWithException() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error               = $mocks['wp_error'];
 
 		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
@@ -63,24 +66,22 @@ class TestHasPageRule extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andThrow( new \Exception() );
-		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
+		$cloudflare_facade_mock->shouldReceive('change_cache_level')->andThrow( new \Exception() );
 
 		$this->assertEquals(
 			new \WP_Error(),
-			$has_page_rule
+			$cloudflare->set_cache_level( 'aggressive' )
 		);
 	}
 
 
 	/**
-	 * Test Cloudflare has page rules with no success.
+	 * Test set cache level with no success.
 	 */
-	public function testHasRuleWithNoSuccess() {
+	public function testSetCacheLevelWithNoSuccess() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error               = $mocks['wp_error'];
 
 		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
@@ -91,53 +92,22 @@ class TestHasPageRule extends TestCase {
 
 		Functions\when( 'wp_sprintf_l' )->justReturn( '' );
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cf_page_rule = json_decode('{"success":false,"errors":[{"code":7003,"message":"Could not route to \/zones\/ZONE_ID, perhaps your object identifier is invalid?"},{"code":7000,"message":"No route for that URI"}],"messages":[],"result":null}');
-		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
-		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
+		$cf_reply   = json_decode('{"success":false,"errors":[{"code":1007,"message":"Invalid value for zone setting cache_level"}],"messages":[],"result":null}');
+		$cloudflare_facade_mock->shouldReceive('change_cache_level')->andReturn( $cf_reply );
 
 		$this->assertEquals(
 			new \WP_Error(),
-			$has_page_rule
+			$cloudflare->set_cache_level( 'aggressive' )
 		);
 	}
 
 	/**
-	 * Test Cloudflare has page rules with success but no page rule.
+	 * Test set cache level with success.
 	 */
-	public function testHasRuleWithSuccessButNoPageRule() {
+	public function testSetCacheLevelWithSuccess() {
 		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
 
 		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error               = $mocks['wp_error'];
-
-		// The Cloudflare constructor run with transient set as WP_Error.
-		Functions\when( 'get_transient' )->justReturn( true );
-		$cloudflare_facade_mock->shouldNotReceive('is_api_keys_valid');
-		Functions\expect( 'set_transient' )->never();
-		Functions\when( 'is_wp_error' )->justReturn( false );
-
-		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
-
-		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cf_page_rule = json_decode('{"result":[{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"bypass"}],"priority":3,"status":"active","created_on":"","modified_on":""},{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":""}],"priority":2,"status":"active","created_on":"","modified_on":""}],"success":true,"errors":[],"messages":[]}');
-		Functions\when( 'wp_json_encode' )->justReturn( json_encode( $cf_page_rule ) );
-		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
-		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
-
-		$this->assertEquals(
-			false,
-			$has_page_rule
-		);
-	}
-
-	/**
-	 * Test Cloudflare has page rules with success and page rule.
-	 */
-	public function testHasRuleWithSuccessAndPageRule() {
-		$mocks = $this->getConstructorMocks( 1,  '',  '', '');
-
-		$cloudflare_facade_mock = $mocks['facade'];
-		$wp_error               = $mocks['wp_error'];
 
 		// The Cloudflare constructor run with transient set as WP_Error.
 		Functions\when( 'get_transient' )->justReturn( true );
@@ -147,14 +117,12 @@ class TestHasPageRule extends TestCase {
 		$cloudflare_facade_mock->shouldReceive('set_api_credentials');
 
 		$cloudflare = new Cloudflare( $mocks['options'], $cloudflare_facade_mock );
-		$cf_page_rule = json_decode('{"result":[{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"bypass"}],"priority":3,"status":"active","created_on":"","modified_on":""},{"id":"","targets":[{"target":"url","constraint":{"operator":"matches","value":""}}],"actions":[{"id":"cache_level","value":"cache_everything"}],"priority":2,"status":"active","created_on":"","modified_on":""}],"success":true,"errors":[],"messages":[]}');
-		Functions\when( 'wp_json_encode' )->justReturn( json_encode( $cf_page_rule ) );
-		$cloudflare_facade_mock->shouldReceive('list_pagerules')->andReturn( $cf_page_rule );
-		$has_page_rule = $cloudflare->has_page_rule( 'cache_everything' );
+		$cf_reply = json_decode('{"result":{"id":"cache_level","value":"aggressive","modified_on":"","editable":true},"success":true,"errors":[],"messages":[]}');
+		$cloudflare_facade_mock->shouldReceive('change_cache_level')->andReturn( $cf_reply );
 
 		$this->assertEquals(
-			true,
-			$has_page_rule
+			'aggressive',
+			$cloudflare->set_cache_level( 'aggressive' )
 		);
 	}
